backend/requirements.txt
alembic==1.4.2
appdirs==1.4.4
atomicwrites==1.4.0
attrs==19.3.0
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
colorama==0.4.3
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
more-itertools==8.4.0
mysqlclient==1.4.6
packaging==20.4
pluggy==0.13.1
py==1.8.2
pycountry==19.8.18
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.17
virtualenv==20.0.25
wcwidth==0.2.5
Werkzeug==1.0.1
